Ride Armor Modularity
wow this took a while!
Ride Armors are now Modular, can have different colors applied to each
part.
Parts will be able to be changed (for now, Wheat turns the part red)
PartsRegistry handles keeping all the parts in order by type then slot,
handles part size and offset.
Model has arms and backpack broken off the torso to be rendered
seperately, will need to fix the rotation point placement when torso
turns.
Ride Armor uses Horse as the base instead of the Boat, tracks the
players facing and moves a bit better, though the rotation is jittery
(it is when riding a horse, but less noticeable).
Added a mech design layout document for how a mech parts set would be
set up if done using JSON, may or may not do that in the future.