Update to Mech Bay
Blocks now properly disappear when multiblock structure is complete.
Extended the back of the mech bay to occupy the corners.
Ride Armor now gets centered and rotated properly when on the mech bay
platform, giving the mech bay a pointer to it (I hope? ).
Description Packet is now handled in TileMultiBlock.
Forgot to upload temp textures for mechbay.