Mech Bay Gui now works
Will change what's the mech bays inventory to what's currently on the
mech.
Can change what's on the mech by opening up the controller and changing
the parts.
Parts will only fit in the appropriate spots, (body in body, left arm in
left arm, left and right are swapped to better reflect the visual model
in the gui).
will have to make sure that arm recipes don't mirror for arms, may need
a work around or something. maybe check if arm is the same on both sides
and allow left/right to be in either slot?
Part registry now handles setting parts to metadata, will be changed to
do nbt data soon for when addons load in different orders or something