Mechbay ready for use!
Can now change parts on the mech, removing the body will break the
entire thing.
If there is no mech in the bay, then you have to place the body part in
first to spawn a new mech.
Water and lava resist should work, but the mech (true to it's
EntityCreature-ness) floats to the top instead of sinking like I want it
to.
Frog Head now has a part, but no model, allows the rider to breathe
underwater.
Debugger now prints to player's chat, because I could.